requirements.txt
PyQt6>=6.4.0
psycopg2-binary>=2.9.5
python-barcode>=0.14.0
Pillow>=9.3.0
requests>=2.28.1
python-dotenv>=0.21.0
bcrypt>=4.0.1
